fix(csr): Set type of node_props to float64 (#235)
`dtype` of `node_props` passed from `csr_to_nbgraph()` to `NBGraph` was
`float32`. This sets the type to explicitly be `np.float64` which
matches the expectation of `csr_spec_float` passed to
`numba.experimental.jitclass()`.

The tests that were previously failing in
[TopoStats](https://github.com/AFM-SPM/TopoStats/) (see #234 for
details) now pass locally. :slightly_smiling_face:

Closes #234

---------

+++ src/skan/csr.py
@@ -207,7 +207,7 @@ def csr_to_nbgraph(csr, node_props=None):
             csr.indices,
             csr.data,
             np.array(csr.shape, dtype=np.int32),
-            node_props,
+            node_props.astype(np.float64),
             )
 
 
@@ -525,7 +525,7 @@ class Skeleton:
         if np.issubdtype(skeleton_image.dtype, np.floating):
             self.pixel_values = skeleton_image[coords]
         elif np.issubdtype(skeleton_image.dtype, np.integer):
-            self.pixel_values = skeleton_image.astype(float)[coords]
+            self.pixel_values = skeleton_image.astype(np.float64)[coords]
         else:
             self.pixel_values = None
         self.graph = graph
